Bugfixes in CodeMirror integration
Fixed a number of bugs in the CodeMirror integration.
 - Proper clean-up of listeners when detach_cm is called
 - Minor typo, where wrong function was called on sharedoc
 - Fixed the generation of sharejs ops. The algorithm in applyToShareJS
was written under the assumption that editorDoc would give you access
to the text BEFORE the change event. This is not the case. Hence the op
is created based on the old text in the sharejs doc now. NB: This is
horribly slow since it requires doing a split on the full text per
event!!